requirements.txt
click==7.1.2
Flask==1.1.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.24.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.13.0
six==1.15.0
Werkzeug==1.0.1
